# server/src/routers/swagger.json
{
...
  "swagger": "2.0",
  "info": {
    "version": "1.0.0",
    "title": "API Documentation for AMS system",
    "description": "List of all APIs for management resources like auditorium, tickets, booking etc..",
    "contact": {
      "name": "Ajay Pipaliya & Jay Patel",
      "email":"apipaliya02@gmail.com"
    },
    "license": {
      "name": "MIT",
      "url": "https://opensource.org/licenses/MIT"
    }
  },
  "basePath": "/",
...
  "produces": ["application/json"],
  "securityDefinitions": {
    "Bearer": {
      "type": "apiKey",
      "name": "Authorization",
      "in": "header"
    }
  },
  "paths": {
    "/users/login": {
      "post": {
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "required": true,
            "description": "Enter valid details to login user",
            "schema": {
              "type": "object",
              "required": ["email", "password"],
              "properties": {
                "email": {
                  "type": "string"
                },
                "password": {
                  "type": "string"
                }
              }
            }
          }
        ],
        "tags": ["General"],
        "summary": "login for all users",
        "responses": {
          "200": {
            "description": "ok"
          },
          "400": {
            "description": "Error while logging into system."
          }
        }
      }
    },
    "/users/signup": {
      "post": {
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "required": true,
            "description": "Enter user's all details for signup",
            "schema": {
              "type": "object",
              "required": ["email", "password", "role", "age", "name"],
              "properties": {
                "name": {
                  "type": "string"
                },
                "age": {
                  "type": "integer"
                },
                "role": {
                  "type": "string"
                },
                "email": {
                  "type": "string"
                },
                "password": {
                  "type": "string"
                }
              }
            }
          }
        ],
        "tags": ["General"],
        "summary": "Signup for all users excepts manager",
        "responses": {
          "201": {
            "description": "User has been created into system."
          },
          "400": {
            "description": "Error while creating user."
          }
        }
      }
    },
...
    "/users/me": {
      "get": {
        "security": [
          {
            "Bearer": []
          }
        ],
        "tags": ["General"],
        "summary": "Get the personal(looged user's) details.",
        "responses": {
          "200": {
            "description": "ok",
            "schema": {
              "$ref": "#/definitions/User"
            }
          },
          "401": {
            "description": "User not authenticated."
          }
        }
      }
    },
...
    "/organizer/auditorium": {
      "get": {
        "parameters": [
          {
            "name": "city",
            "in": "query",
            "type": "string",
            "description": "get all auditorium list using filter by city"
          }
        ],
        "tags": ["Organizer"],
        "summary": "Get the all auditorium list for organizer",
        "security": [
          {
            "Bearer": []
          }
        ],
        "responses": {
          "200": {
            "description": "Ok"
          },
          "401": {
            "description": "Please authenticate befor sending request"
          },
          "400": {
            "description": "Unauthenticated person to view this details"
          }
        }
      }
    },
    "/organizer/getAvailableSlots": {
      "get": {
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "required": true,
            "description": "Enter auditorium id and date details ",
            "schema": {
              "type": "object",
              "required": ["auditorium_id", "date"],
              "properties": {
                "auditorium_id": {
                  "type": "string"
                },
                "date": {
                  "type": "string"
                }
              }
            }
          }
        ],
        "tags": ["Organizer"],
        "summary": "Get the all events list of auditorium for event booking",
        "security": [
          {
            "Bearer": []
          }
        ],
        "responses": {
          "200": {
            "description": "Ok"
          },
          "401": {
            "description": "Please authenticate befor sending request"
          },
          "400": {
            "description": "Unauthenticated person to view this details"
          }
        }
      }
    },
    "/organizer/bookAuditorium": {
      "post": {
        "security": [
          {
            "Bearer": []
          }
        ],
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "required": true,
            "description": "Enter all details of event for auditorium booking",
            "schema": {
              "type": "object",
              "required": [
                "timeslots",
                "event_date",
                "auditorium_id",
                "description",
                "event_name",
                "category",
                "event_time",
                "ticket_price"
              ],
              "properties": {
                "timeslots": {
                  "type": "array"
                },
                "event_date": {
                  "type": "string"
                },
                "event_time": {
                  "type": "string"
                },
                "ticket_price": {
                  "type": "number"
                },
                "auditorium_id": {
                  "type": "string"
                },
                "description": {
                  "type": "string"
                },
                "event_name": {
                  "type": "string"
                },
                "category": {
                  "type": "string"
                }
              }
            }
          }
        ],
        "tags": ["Organizer"],
        "summary": "Book auditorium for event",
        "responses": {
          "200": {
            "description": "Ok"
          },
          "401": {
            "description": "Please authenticate befor sending request"
          },
          "400": {
            "description": "Unauthenticated person to view this details"
          }
        }
      }
    },
    "/organizer/audiBookingPayment/:status": {
      "post": {
        "security": [
          {
            "Bearer": []
          }
        ],
        "parameters": [
          {
            "name": "status",
            "in": "query",
            "type": "string",
            "description": "Enter status value True if want to make payment success otherwise enter False"
          },
          {
            "name": "body",
            "in": "body",
            "required": true,
            "description": "Enter event_id and amount to make payment",
            "schema": {
              "type": "object",
              "required": ["event_id"],
              "properties": {
                "amount": {
                  "type": "number"
                },
                "event_id": {
                  "type": "string"
                }
              }
            }
          }
        ],
        "tags": ["Organizer"],
        "summary": "Check payment status for event booking",
        "responses": {
          "200": {
            "description": "Ok"
          },
          "401": {
            "description": "Please authenticate befor sending request"
          },
          "400": {
            "description": "Unauthenticated person to view this details"
          }
        }
      }
    }
  },
  "definitions": {
    "User": {
      "required": [
        "_id",
        "name",
        "email",
        "age",
        "role",
        "verificationStatus",
        "password",
        "tokens"
      ],
      "properties": {
        "_id": {
          "type": "string",
          "uniqueItems": true
        },
        "name": {
          "type": "string"
        },
        "email": {
          "type": "string",
          "uniqueItems": true
        },
        "age": {
          "type": "number"
        },
        "role": {
          "type": "string"
        },
        "password": {
          "type": "string"
        }
      }
    },
...
  }
}